update: カスタムフックに切り出し
