+++ frontend/src/_components/features/Login/Body/index.tsx
@@ -1,43 +1,16 @@
 'use client'
 
 import { EMAIL_REGEX } from '@/constants/regexes'
-import { useAuth } from '@/hooks/use-Auth'
 import { Button } from '@/_components/common/button'
 import { InputField } from '@/_components/common/input-field'
 import { Box, Stack } from '@mui/material'
-import { useRouter } from 'next/navigation'
-import { FieldValues, SubmitHandler, useForm } from 'react-hook-form'
-import { toast } from 'react-toastify'
+import { useInputForm } from './logics/use-input-form'
 
 export const LoginBody = () => {
-  const router = useRouter()
-  const { login, isLoading } = useAuth()
-
-  const { control, handleSubmit } = useForm({
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-    mode: 'onBlur',
-  })
-
-  const onSubmitLogin: SubmitHandler<FieldValues> = (data) => {
-    login({
-      data,
-      onSuccess: () => {
-        router.push(process.env.NEXT_PUBLIC_FRONT_END || '')
-        toast.success('ログインに成功しました。')
-      },
-      onError: () => toast.error('ログインに失敗しました。'),
-    })
-  }
-
+  const { control, isLoading, handleSubmit } = useInputForm()
+  
   return (
-    <Box
-      component={'form'}
-      width={'100%'}
-      onSubmit={handleSubmit(onSubmitLogin)}
-    >
+    <Box component={'form'} width={'100%'} onSubmit={handleSubmit}>
       <Stack direction={'column'} spacing={2} padding={'16px'}>
         <InputField
           control={control}
