+++ frontend/src/_components/features/Todo/Item/index.tsx
@@ -1,36 +1,26 @@
 'use client'
 
-import { todoAtom } from '@/store'
 import { TodoType } from '@/types/todo'
 import { Button } from '@/_components/common/button'
 import { ListItem, Stack, Typography } from '@mui/material'
-import { useSetAtom } from 'jotai'
+import { useTodo } from './logics/use-todo'
 
 type Props = {
   item: TodoType
-  refetch: () => void
   onEditModalOpen: () => void
   onDeleteConfirmModalOpen: () => void
 }
 
 export const TodoItem = ({
   item,
-  refetch,
   onEditModalOpen,
   onDeleteConfirmModalOpen,
 }: Props) => {
-  const setTodo = useSetAtom(todoAtom)
+  const { editTodo, deleteTodo } = useTodo(
+    onEditModalOpen,
+    onDeleteConfirmModalOpen
+  )
 
-  const onClickEvents = {
-    editTodo: (todo: TodoType) => {
-      setTodo({ id: todo.id, detail: todo.detail, title: todo.title })
-      onEditModalOpen()
-    },
-    deleteTodo: (todo: TodoType) => {
-      setTodo({ id: todo.id, detail: todo.detail, title: todo.title })
-      onDeleteConfirmModalOpen()
-    },
-  }
   return (
     <ListItem
       disableGutters
@@ -43,15 +33,11 @@ export const TodoItem = ({
     >
       <Typography fontSize={'20px'}>{item.title}</Typography>
       <Stack direction={'row'} gap={1}>
-        <Button
-          text={'編集'}
-          fullWidth
-          onClick={() => onClickEvents.editTodo(item)}
-        />
+        <Button text={'編集'} fullWidth onClick={() => editTodo(item)} />
         <Button
           text={'削除'}
           color={'error'}
-          onClick={() => onClickEvents.deleteTodo(item)}
+          onClick={() => deleteTodo(item)}
         />
       </Stack>
     </ListItem>
