+++ frontend/src/_components/features/Todo/add-area/index.tsx
@@ -1,56 +1,12 @@
 'use client'
 
-import { useFetchTodoList } from '@/hooks/use-fetch-todo-list'
-import { usePostCreateTodo } from '@/hooks/use-post-create-todo'
-import { authUserAtom } from '@/store'
-import { TodoCreateRequest } from '@/types/todo'
 import { Button } from '@/_components/common/button'
 import { InputField } from '@/_components/common/input-field'
 import { Box } from '@mui/material'
-import { useAtomValue } from 'jotai'
-import { useRef, useState } from 'react'
-import { SubmitHandler, useForm } from 'react-hook-form'
+import { useInputForm } from './logics/use-input-form'
 
 export const TodoAddArea = () => {
-  const { refetch } = useFetchTodoList()
-  const { doPost } = usePostCreateTodo()
-  const authUser = useAtomValue(authUserAtom)
-  const [isCreateLoading, setIsCreateLoading] = useState(false)
-  const submitProcessing = useRef(false)
-  const { control, handleSubmit, reset } = useForm<TodoCreateRequest>({
-    defaultValues: {
-      title: '',
-      detail: '',
-    },
-    mode: 'onBlur',
-  })
-
-  const handleOnSubmitCreateTodo: SubmitHandler<TodoCreateRequest> = (data) => {
-    if (submitProcessing.current) return
-
-    const postData = {
-      ...data,
-      user_id: authUser?.id,
-    }
-
-    submitProcessing.current = true
-    setIsCreateLoading(true)
-
-    doPost({
-      data: postData,
-      onSuccess: () => {
-        refetch()
-        reset()
-      },
-      onError: () => {
-        throw new Error('作成できませんでした')
-      },
-      onFinally: () => {
-        submitProcessing.current = false
-        setIsCreateLoading(false)
-      },
-    })
-  }
+  const {control, isCreateLoading, handleSubmit} = useInputForm()  
 
   return (
     <Box
@@ -58,7 +14,7 @@ export const TodoAddArea = () => {
       marginTop={'16px'}
       display={'grid'}
       gap={'16px'}
-      onSubmit={handleSubmit(handleOnSubmitCreateTodo)}
+      onSubmit={handleSubmit}
     >
       <Box display={'grid'} gap={'16px'}>
         <InputField
