+++ frontend/src/app/page.tsx
@@ -1,35 +1,12 @@
-'use client'
-
 import { HeaderNavigation } from '@/_components/common/header-navigation'
 
 import { TodoList } from '@/_components/features/todo/list'
 
-import { useAuth } from '@/hooks/use-Auth'
 import AppSuspense from '@/_components/common/app-suspense'
 import { TodoAddArea } from '@/_components/features/todo/add-area'
 import { Box, Container, Typography } from '@mui/material'
-import { useEffect } from 'react'
-
-/**
- * NOTE:
- * dynamic import を活用することで非同期にモジュールを読み込むことができる。（返り値はPromise）
- * オプションに {ssr: false} を指定することで、クライアントサイドでレンダリングすることを可能とする。
- * ページ読み込み時に必要なJSファイルを削減できるため、ページ表示速度改善にもつながる。
- */
 
 export default function Page() {
-  const { fetchMe } = useAuth()
-
-  // HACK:
-  // ログイン時に fetchMe を叩いてるため、結果的に fetchMe を複数回実行していることになる。
-  // こういう設計を許していると、リクエスト数が無駄に跳ね上がり、サーバー側が逼迫する原因になりかねないので、
-  // もっといいやりかたがないか検討する。
-  useEffect(() => {
-    fetchMe()
-    // NOTE: 初回レンダリング時にのみ fetchMe を実行したいため、依存配列は空のままにする
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
   return (
     <>
       <Container sx={{ marginTop: '56px', paddingTop: '20px' }}>
